requirements.txt
python-telegram-bot==21.6
matplotlib
requests
beautifulsoup4
nest_asyncio
